feature/W3RCT-4 Code refactoring and splitting into components
