+++ src/components/Footer.scss
@@ -1,0 +1,34 @@
+@use "src/globalStyleVariables";
+
+.footer {
+  background-color: #231f20;
+  height:fit-content;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  hr {
+    background-color: #D3D3D3;
+    color: #D3D3D3;
+    border-color: #D3D3D3;
+    height: 3px;
+    width: 60%;
+  }
+  h1{
+    color: #F9F6EE;
+    font-size: 24px;
+    margin-bottom: 5px;
+
+  }
+  h3{
+    margin-top: 5px;
+    margin-bottom: 5px;
+    font-size: 15px;
+    color: #F9F6EE;
+    @media screen and (min-width: globalStyleVariables.$small){
+      text-align: center;
+    }
+
+  }
+}
