+++ src/components/LanguageBar.scss
@@ -1,0 +1,30 @@
+@use "src/globalStyleVariables";
+
+.languageHeader{
+  background-color: #2b2828;
+  height: 30px;
+  max-width: available;
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+
+  .btn{
+    cursor: pointer;
+    border: 0 solid;
+    background-color: transparent;
+    height: 25px;
+    width: 50px;
+    color: #FFFFFF;
+    font-size: 15px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 10px;
+  }
+
+  .language-flag{
+    width: 23px;
+    height: 13px;
+    margin-right: 5px;
+  }
+}
