+++ src/components/pages/Home/HomePageContent.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+import "./HomePageContent.scss";
+// import strings from "../../../config/strings";
+
+// const { homepage: homepageStrings } = strings;
+
+const HomePageContent = () => {
+  return (
+    <>
+      <div className="content">
+        <h1>Content</h1>
+      </div>
+    </>
+  );
+};
+
+export default HomePageContent;
