+++ src/components/pages/Home/HomePageContent.scss
@@ -1,0 +1,6 @@
+@use "src/globalStyleVariables";
+
+.content {
+  height: 500px;
+  background-color: #D3D3D3;
+}
