+++ src/config/strings/general.js
@@ -1,0 +1,25 @@
+// eslint-disable-next-line import/no-anonymous-default-export
+export default {
+  en: {
+    general: {
+      romanianLanguageButton: "RO",
+      englishLanguageButton: "EN",
+      header: {
+        menuButtons: {
+          home: "Home",
+          news: "News",
+          schedule: "Schedule",
+          contact: "Contact",
+          aboutUs: "About Us",
+        },
+      },
+      footer: {
+        slogan: "The voice of the students.",
+        phoneNumber: "0712 345 678",
+        email: "unitbvradiocampus@gmail.com",
+        siteCreator:
+          "© 2022 Radio Campus Transilvania (RCT). Created By Alex Gheorghe.",
+      },
+    },
+  },
+};
