+++ src/config/strings/homepage.js
@@ -3,20 +3,7 @@ export default {
   en: {
     homepage: {
       title: "Home",
-      menuButtons: {
-        home: "Home",
-        news: "News",
-        schedule: "Schedule",
-        contact: "Contact",
-        aboutUs: "About Us",
-      },
-      footerDetails: {
-        slogan: "The voice of the students.",
-        phoneNumber: "0712 345 678",
-        email: "unitbvradiocampus@gmail.com",
-        siteCreator:
-          "© 2022 Radio Campus Transilvania (RCT). Created By Alex Gheorghe.",
-      },
+      content: {},
     },
   },
 };
